feat(core/rust): introduce f32 angles for sector drawing
[no changelog]

+++ core/embed/rust/src/ui/shape/circle.rs
@@ -15,8 +15,8 @@ pub struct Circle {
     bg_color: Option<Color>,
     thickness: i16,
     alpha: u8,
-    start_angle: Option<i16>,
-    end_angle: Option<i16>,
+    start_angle: Option<f32>,
+    end_angle: Option<f32>,
 }
 
 impl Circle {
@@ -55,14 +55,14 @@ impl Circle {
         Self { alpha, ..self }
     }
 
-    pub fn with_start_angle(self, from_angle: i16) -> Self {
+    pub fn with_start_angle(self, from_angle: f32) -> Self {
         Self {
             start_angle: Some(from_angle),
             ..self
         }
     }
 
-    pub fn with_end_angle(self, to_angle: i16) -> Self {
+    pub fn with_end_angle(self, to_angle: f32) -> Self {
         Self {
             end_angle: Some(to_angle),
             ..self
@@ -119,8 +119,8 @@ impl Shape<'_> for Circle {
                 }
             }
         } else {
-            let start = self.start_angle.unwrap_or(0);
-            let end = self.end_angle.unwrap_or(360);
+            let start = self.start_angle.unwrap_or(0.0);
+            let end = self.end_angle.unwrap_or(360.0);
 
             if let Some(color) = self.fg_color {
                 if th > 0 {
